Fix compilation of all-benchmarks
